server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
motor>=3.3.1
pyjwt>=2.8.0
python-dotenv>=1.0.1
